requirements.txt
Django==2.1.4
djangorestframework==3.9.0
django-cors-headers==2.4.0
coverage-4.5.2